Status typedefs
