+++ typedefs.js
@@ -1,6 +1,7 @@
 const { gql } = require("apollo-server-express");
+const { typeDefs: Status } = require("./resolvers/Status");
 
-const typeDefs = gql`
+const base = gql`
   type PageInfo {
     hasNextPage: Boolean!
     "Total of nodes returned from query."
@@ -61,11 +62,6 @@ const typeDefs = gql`
     nationality: String
     "http://en.wikipedia.org/wiki/Fernando_Alonso"
     driverUrl: String
-  }
-
-  type DriverRaceTime {
-    millis: String
-    time: String
   }
 
   type RaceCompetitorFastestLapAverageSpeed {
@@ -301,42 +297,6 @@ const typeDefs = gql`
     pageInput: PageInput!
   }
 
-  type StatusCode {
-    "e.g. '1'"
-    id: ID
-    "count '13'"
-    count: Int
-    "Finished"
-    name: String
-  }
-
-  type StatusCodesReport {
-    nodes: [StatusCode]
-    pageInfo: PageInfo!
-  }
-
-  input StatusCodesSearchInput {
-    "Finishes in a season. e.g. current, 2021, 2008"
-    season: String
-    "Finishes in a certain round. e.g. last, 1, 12"
-    round: String
-    "Finishes for a particular constructor. e.g. 'mclaren'"
-    constructors: String
-    "Finishes for a particular driver. e.g. 'alonso'"
-    drivers: String
-    "Finishes at a particular circuit. e.g. 'monza'"
-    circuits: String
-    "Finishes for a driver that started a race in a specified position. e.g. '1'"
-    grid: String
-    "Specify a particular finish code ID"
-    status: String
-  }
-
-  input StatusCodesInput {
-    where: StatusCodesSearchInput
-    pageInput: PageInput!
-  }
-
   type Query {
     Constructor(id: ID): Constructor
     Constructors(input: ConstructorsInput): ConstructorsReport
@@ -351,5 +311,7 @@ const typeDefs = gql`
     Status(input: StatusCodesInput): StatusCodesReport
   }
 `;
+
+const typeDefs = [base, Status];
 
 module.exports = { typeDefs };
